Made the terminal (with commands). Next is to add the Market to it.

+++ Market.h
@@ -22,8 +22,8 @@ private:
   Market();
 
   const vector<Asset>& getAssets() const;
-  bool buy(Player& player, double assetIndex, double amount);
-  bool sell(Player& player, double assetIndex, double amount);
+  bool buy(Player& player, int assetIndex, double amount);
+  bool sell(Player& player, int  assetIndex, double amount);
 
   void updatePrices();
 };
